5 Summer 2023 (May 2023 - August 2023)/SENG 275 A01 B03/1 Assignments and Labs/0 GitLab/2 Assignments/A1/lib/src/main/java/assignment1/Triangle.java
package assignment1;

public class Triangle {
    public enum TriangleType {
        EQUILATERAL,
        ISOSCELES,
        SCALENE,
        INVALID
    }

    public static TriangleType categorize(int x, int y, int z) {
        if (x > 0 && y > 0 && z > 0 && x + y >= z && x + z >= y && y + z >= x) {
            if (x == y && y == z) {
                return TriangleType.EQUILATERAL;
            } else if (x == y || y == z || x == z) {
                return TriangleType.ISOSCELES;
            } else {
                return TriangleType.SCALENE;
            }
        } else {
            return TriangleType.INVALID;
        }
    }

    public static void main(String[] args) {
        int x = 5;
        int y = 5;
        int z = 5;

        TriangleType type = Triangle.categorize(x, y, z);

        System.out.println("Triangle type: " + type);
    }
}
